dissectors: new dissector for the MDB protocol
The MDB (Multi-Drop Bus) protocol is used inside a vending machine. MDB
defines the communication between the main control board (VMC = Vending
Machine Controller) and peripheral components, e.g. a payment terminal
or a bill validator.

The protocol specification is available from
https://namanow.org/nama-releases-mdb-version-4-3/

The pcap input format for this dissector is documented at
https://www.kaiser.cx/pcap-mdb.html

We are in the process of requesting an official DLT for MDB.
For now, the dissector can be mapped to a User-DLT for testing.