Telnet: Dissect VMware Serial Port Proxy extension.
The VMware Serial Port Proxy Telnet extensions are used for managing the
connection between a network serial port for a virtual machine running on
VMware ESXi and a virtual Serial Port Concentrator (vSPC).  Most
notably, they are used to manage the handoff of a network serial port
connection from one ESXi host to another during a vMotion (virtual machine live
migration) event, with the aim of ensuring that the virtual machine's serial
port will not lose connectivity at any point during the migration.

The protocol is specified in "Using a Proxy with Virtual Serial Ports", which
may be accessed at
https://developer.vmware.com/docs/11763/using-a-proxy-with-virtual-serial-ports

This protocol usually runs over TCP port 13370.  Use "Decode As..." or specify
"-d tcp.port==13370,telnet" on the command-line.

An example of a vSPC implementation can be found at
https://github.com/pbhenson/vSPC.py

This change implements the bulk of the protocol which does not require tracking
the vSPC vMotion conversation -- the vSPC vMotion conversation itself will span
two Telnet conversations, one from the source host to the vSPC, and the other
from the destination host to the vSPC.

Most notably, this precludes (for now) dissecting the data carried with the
VMOTION-GOAHEAD and VMOTION-PEER commands, where it is not possible to
determine where the "sequence" field finishes and "secret" starts unless the
packet can be matched up with the prior VMOTION-BEGIN message containing the
"sequence" field alone.