epan: Fix up --disable-all-protocols
Commit af0691342b7e6bc2ef81a8430ae00680223337bc and
commit 80ae370811595f294c65cb71e1d213d5b49e33fd both added
proto_disable_all.

Take the one from the 80ae370811595f294c65cb71e1d213d5b49e33fd,
which made changes to allow protocols to be enabled by
proto_enable_proto_by_name() even if they are not disabled by
default, which allows --only-protocols to work without changing
the disabled-by-default setting of a protocol.

We don't want to change the disabled by default setting, because
that causes the disabled status to be persistent even when
changing Configuration Profiles with the GUI, instead of being
reset. In general, command line options that temporarily override
the Configuration Profile at startup are cleared when switching
to a new profile via the profile, and we don't want to make an
exception.