BT-DHT: Fix infinite loop in malformed error packets
If the int or string that is part of the error returns 0 (indicating
failure) then return zero. Commit 6f30bc471346e3e2ce3df1e13d873669fa94c951
fixed an off-by-one error in returning the end offset of an error, but
that caused 1 instead of 0 to be returned from dissect_bt_dht_error,
leading to stack overflows and infinite loops in the case of a badly
formatted error.

Fix #19277.