editcap: Add --discard-packet-comments option and fix -a option
Added new `editcap` option `--discard-packet-comments` to discard all
packet comments when editing a pcap file and writing a new one.  This
behaves the same way as the existing `--discard-capture-comment`
option only it discards packet comments and not capture file comments.
Packet comments added with `-a` on the same command line are not
discarded.

Also, fix the existing `-a` option to work the way the documentation,
which says it should "Add or replace comment for given frame number",
describes it.  Namely, any existing comments for the packet are now
removed before the comment specified by the `-a` option is added.