dissect_opts: use a better name for a routine.
Yes, *one* of the things the routine does is to split a comma-separated
list into individual items, but the *main* thing it does is process all
items in an slist of entries that are strings that are comma-separated
lists of items by passing them to a callback routine.

Rename the routine to describe what it does, namely process an enable or
disable list in the aforementioned format, calling a routine that
enables or disables the item.