Qt: Close console dialog when reloading Lua
Fixes a crash using the console because the dialog holds a pointer
to the Lua state that gets invalidated by the reload.

Forcing the dialog to close drops the invalid reference and avoids
the crash and all the other attending state problems with the reload.

In theory there might be console types other than Lua so we may
want to fix the code to close only Lua type dialogs if only Lua
plugins are reloaded.