Qt: Remove tail update timer from packet list
We have a preference that controls how often we get capture updates,
so there's no reason (nor CPU savings) to have a separate timer
checking for when to auto scroll. We can just scroll to the bottom
whenever we insert new rows (which doesn't happen any faster than
prefs.capture_update_interval, which the user can change.)

This makes the auto scrolling smoother, instead of having a gap where
rows have been inserted but not scrolled yet. If the scrolling is too
fast, a user can change the preference.