manuf: Print short name with tshark -G
Add the short name (used by Wireshark for MAC address name resolution)
to the output as another column. This matches the historical
format for this text file, before it was turned into static
binary data.

The use case for still parsing an optional external manuf text file
is to allow customization of the compiled in data, for any reason.

In that case it is useful to be able to run something like:

    tshark -G manuf | grep MyVendorName >> ~/.config/wireshark/manuf

And then edit the newly added short name entries to one's satisfaction.