ICMP: Improve heuristic for data time
When determining if the first few bytes of a ICMP Echo or
Echo Reply look like a timestamp, take the smaller difference
from the packet time instead of just taking the Big Endian time
if it's less than our maximum delta.

Also, test for timevals using 64 bit time_t, since that's common
now. Assume that suseconds_t is 64 bit if time_t is, even though
it doesn't have to be (do any 32 bit systems do that?)

Assume that if something looks like a LE timeval with a 64 bit time_t
that it is one, rather than being a LE timeval with a 32 bit time_t
and 0 fractional seconds. (Otherwise we'd make the wrong heuristic
decision with clock skew that makes the ping timestamp in the future.)

Fix #19283