tshark: Restore deprecated default column field titles
Restore the ability in tshark -e to filter based on column titles,
at least for the seven default columns.  Internally alias the fields
based on their titles to the column type based fields, e.g.
"_ws.col.Info" to "_ws.col.info"

This may be removed in the future, but is useful to have as there
are tools and scripts that use the old beahvior (and there are
supports versions, 3.6.x and 4.0.x, that only support the old
behavior.)