Keep Lua console open and fix crash on Lua reload
Use global variables to save the state needed to restore the
print-to-console Lua function when the Lua state is destroyed
and re-created (reloaded).

Do not pass a lua_State pointer to the dialog, for the same reasons;
it is already a global variable, use that instead.

Fixes a crash that was easily reproduced on macOS, but probably
happens elsewhere too, where the saved state became inconsistent on
reload because the dialog wasn't re-registered. (It still isn't,
but now we are handling it with global variables on the Lua side).

Do not close the console dialog when Lua is reloaded. It is no
longer necessary and improves the usability of the dialog.