Qt: Make IOConsoleDialog a single instance
When opening a console dialog, make it reuse an existing instance
if it is already open. This is required because the output of
the Lua code currently only goes into a single GUI dialog
(the last one open). And in general it feels correct.

This depends on the QAction holding the pointer reference,
which seems reasonable because it is a dynamic menu action.