Add some benchmarking timers to tshark
Add timing measurements for tshark when dissecting
and filtering a capture file.

The output is in JSON. The "elapsed" member is the total
elapsed time for all passes. After that there is
an array with an object for each pass. We only
have two passes at most currently. The single
pass dissection (the default) generates
an array with one element.

Currently there is a counter for total time
for the pass, and time elapsed in dissection,
applying the read filter and applying the
display filter. If any of these is not
active the respective entry contains zero.

All values are in milliseconds.

Add a tshark option to print the timing info.

Example:

    $ tshark -r ~/captures/http.pcapng --print-timers -2 -Y 'http && frame.len > 1' > /dev/null
    {
      "version": "v4.1.1rc0-30-gb7e187fe2993",
      "path": "/home/jpv/captures/http.pcapng",
      "filter": "http && frame.len > 1",
      "time_unit": "millisecond",
      "elapsed": 22981243,
      "dfilter_expand": 3,
      "dfilter_compile": 35,
      [
        {
          "elapsed": 11022013,
          "dissect": 10580266,
          "display_filter": 24567,
          "read_filter": 0
        },
        {
          "elapsed": 11959230,
          "dissect": 11454227,
          "display_filter": 21052,
          "read_filter": 0
        }
      ]
    }