CAMEL: Separate time and timezone, check for truncation
ENC_BCD_DIGITS_0_9 truncates the returned string if a
nibble with 0xf is found, so check for that and don't
read past the end of the string. Go ahead and check for
other non-digits as well.

The timezone is not exactly BCD encoded, unlike the time
digits, so handle it separately.

GSM SMS uses the same encoding, so possibly there should be
a special time encoding to handle this and turn it into a
FT_ABSOLUTE_TIME.

Fix #19298