addr_resolv: Don't leak tables on every init dissection
Most (all but the entries manually entered through the GUI) address
resolutions are reset upon init_dissection(), because they have
entries from files contained in configuration profiles and thus
need to be reset when the profile changes. (This is also called
upon loading a new file, or with the -M option to tshark.)

That means that if the tables and their entries are in epan scope,
they will not be freed until the program is exited, which means
that excessive memory is consumed.

Create a new wmem allocation scope and use it for the address
tables and entries. (Some of these maps could be made autoreset with
epan scope and the new scope.)

This is similar to file scope, but some lookups need to be available
before file scope is entered; specifically, the enterprises lookups
are used by the RADIUS dissector for turning vendor numbers into
strings during AVP registration, when happens when the dissector
is registered.

Do NOT use the new scope for the manually entered lookups via
"Edit Resolved Name," because those are not freed upon starting
a new dissection (should they be?)

Fix #19309