log3gpp: remove #if 0'ed and commented-out stub and unmaintained code.
wtap_register_phonelog() is unmaintained and is already out-of-date;
future API changes will render it further out-of-date (and no rationale
is given for passing TRUE as the "add this first" flag to
wtap_register_open_info()).

The code to support packet writing is all stubs; if it's useful to
support writing files in this format, write the code to do so and submit
it.  The stubs in question represent a small fraction of the code
necessary to support this, and don't add much.