dfilter: Extend support for timezones in filter expressions
Extends the dfilter syntax with more the timezones supported
by ws_strptime().

We remove the handling of tm_gmtoff and tm_zone in struct tm,
because it is more trouble than it is worth due to platform
incompatibilites, and instead use an extra arguments to pass
these two non-standard fields.

Remove a test that is no longer relevant (we do not need to
assert that WET in particular is not supported).