dfilter: Fix printing of nanosecond fractions of time.
The code is incorrectly also omitting zeros that are significant
from the representation of fractional values of absolute time
in display filters (with FTREPR_DFILTER).

Instead of fixing the representation just remove the buggy code and
include all 9 digits of fractional seconds, including trailing zeros
to the right of the decimal point, which in this case can be inferred
to indicate the precision (to the nanosecond).