reassemble: Set overlap flags instead of throwing exception
Have the byte offset based fragment_add() routines work like
the fragment sequence number based _seq() routines and set
overlap flags instead of throwing a ReassemblyError for a new
fragment after the reassembly is finished. It's not that
unusual to have a retransmission in such a case, and adding
a Malformed expert info seems worse than linking to the frame
number of the original reassembly.

Change the reassemble test to account for that.

Add some comments to the header file for the reassembly functions
about when to use each function and their respective limitations.