Test: Change the capture used for expert info testing
The http2-data-reassembly.pcap has Errors, Warnings, Notes,
and Chats when TCP checksum verification is turned on, which
allows for somewhat greater test coverage here.

In the future, we might want to make fragment_add set the
overlap flags instead of throwing a ReassemblyError for
retransmissions after a ressembly is complete, which would
make the current test fail.